Flat File Creator
This tool allows users to create flatfile using any combination of tabs in the active workbook. Just select the tabs you want to copy data from and then select the range, CaSES will take care of the rest. The tool will output a formatted data table and save a ton of time.